requirements.txt
pytest-bdd==7.0.1
pytest-playwright==0.4.4
pytest
playwright
python-dotenv
pytest-asyncio

-e .